Create a pointer from `Prop` to `Node`.
When generating code to call processes from HOC or Python one difficulty
can be obtaining the voltage. The voltage is associated with the Node.
In the generated code we only have access to the Prop.

We don't have access to the voltage via `Memb_list::nodeindices`. When
generating code for NRN with NMODL, we don't populate what CoreNEURON
calls `v_unused`.